chore: extract loadbalancer, network, crashdup and process from firecracker
Second part of refactoring to split common logic for VM provisioners
from Firecracker provisioner.
