+++ internal/pkg/provision/providers/vm/crashdump.go
@@ -2,7 +2,7 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
-package firecracker
+package vm
 
 import (
 	"context"
@@ -13,13 +13,12 @@ import (
 	"strings"
 
 	"github.com/talos-systems/talos/internal/pkg/provision"
-	"github.com/talos-systems/talos/internal/pkg/provision/providers/vm"
 	"github.com/talos-systems/talos/internal/pkg/tail"
 )
 
 // CrashDump produces debug information to help with debugging failures.
-func (p *provisioner) CrashDump(ctx context.Context, cluster provision.Cluster, out io.Writer) {
-	state, ok := cluster.(*vm.State)
+func (p *Provisioner) CrashDump(ctx context.Context, cluster provision.Cluster, out io.Writer) {
+	state, ok := cluster.(*State)
 	if !ok {
 		fmt.Fprintf(out, "error inspecting firecracker state, %#+v\n", cluster)
 		return
